ChefSentryDocker/requirements.txt
pyserial
flask
fastai
opencv-python

